Fix UnicodeDecodeError occurs some locales on Python 3
On Python 3, if encoding is not set when text file opens,
locale.getpreferredencoding(False) is used as encoding.
And when encoding of text data is different from
locale.getpreferredencoding(False), UnicodeDecodeError may
occur while reading.

Therefore, this commit detects encoding of text data by
chardet.

Detected encoding is used if chardet can detect it. But
if chardet cannot detect encoding, default encoding of
open() on Python 3 is used.